FloatingPoint: Fixed errorprone complaint.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fix_floatingpoint_calculator@46190 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/floatingpoint/test/CFloatUnitTest.java
@@ -185,7 +185,7 @@ public abstract class CFloatUnitTest {
   }
 
   private int calculateExpWidth(BinaryMathContext pFormat) {
-    return (int) Math.ceil(lb(2 * pFormat.maxExponent + 1));
+    return (int) Math.ceil(lb((double) (2 * pFormat.maxExponent + 1)));
   }
 
   private String toBits(BigFloat value) {
